requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx[http2]==0.25.2
python-dotenv==1.0.0
PyJWT[crypto]==2.8.0
python-multipart==0.0.6
firebase-admin==6.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
schedule==1.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0